+++ src/Company/Settings/CompanySettingsInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Copyright © MB Arbatos Klubas. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace FreshAdvance\ElectronicInvoice\Company\Settings;
+
+interface CompanySettingsInterface
+{
+    public function getRegistryNote(): string;
+}