+++ views/admin_twig/de/module_options.php
@@ -14,4 +14,8 @@ $aLang = [
     'SHOP_MODULE_GROUP_fa_electronic_invoice_main' => 'Einstellungen',
     'SHOP_MODULE_fa_electronic_invoice_shop_country_iso' => 'Länder ISO2 des Verkäufers',
     'HELP_SHOP_MODULE_fa_electronic_invoice_shop_country_iso' => 'Wie in den allgemeinen Shop-Einstellungen ermöglicht die Ländereingabe beliebigen Text. Diese Einstellung stellt sicher, dass wir die korrekte ISO2 für die Rechnung auf der Verkäuferseite haben.',
+
+    'SHOP_MODULE_GROUP_fa_electronic_invoice_company' => 'Unternehmensinformationen',
+    'SHOP_MODULE_fa_electronic_invoice_registry_note' => 'Registervermerk',
+    'HELP_SHOP_MODULE_fa_electronic_invoice_registry_note' => 'Geben Sie alle relevanten Unternehmensregisterinformationen an, wie z.B. Firmenname, Adresse, Handelsregister und Gericht, Vertretungsdetails, Geschäftsführer, Handelsregister der Komplementärin und USt-ID.<br>Beispiel:<br>ACME GmbH, Musterstraße 12, 12345 Berlin<br>Handelsregister: Amtsgericht Berlin, HRB 987654<br>Vertreten durch den Geschäftsführer: Max Mustermann<br>Handelsregister der Komplementärin: Amtsgericht Berlin, HRB 123456<br>USt-ID: DE123456789',
 ];