+++ views/admin_twig/en/module_options.php
@@ -14,4 +14,8 @@ $aLang = [
     'SHOP_MODULE_GROUP_fa_electronic_invoice_main' => 'Settings',
     'SHOP_MODULE_fa_electronic_invoice_shop_country_iso' => 'Seller country ISO2',
     'HELP_SHOP_MODULE_fa_electronic_invoice_shop_country_iso' => 'As in shop main settings, the country input allows anything as text, this setting ensures we have correct ISO2 for the invoice on the Seller side.',
+
+    'SHOP_MODULE_GROUP_fa_electronic_invoice_company' => 'Company Information',
+    'SHOP_MODULE_fa_electronic_invoice_registry_note' => 'Registry Note',
+    'HELP_SHOP_MODULE_fa_electronic_invoice_registry_note' => 'Enter all relevant company registry information, such as company name, address, commercial register and court, representation details, managing director, commercial register of the general partner, and VAT ID (USt-ID).<br>Example:<br>ACME GmbH, Musterstraße 12, 12345 Berlin<br>Handelsregister: Amtsgericht Berlin, HRB 987654<br>Vertreten durch den Geschäftsführer: Max Mustermann<br>Handelsregister der Komplementärin: Amtsgericht Berlin, HRB 123456<br>USt-ID: DE123456789',
 ];